+++ src/Type/Definition/UnionType.php
@@ -110,20 +110,10 @@ class UnionType extends Type implements AbstractType, OutputType, CompositeType,
         return $this->types;
     }
 
-    /**
-     * Resolves concrete ObjectType for given object value
-     *
-     * @param object $objectValue
-     * @param mixed  $context
-     *
-     * @return callable|null
-     */
     public function resolveType($objectValue, $context, ResolveInfo $info)
     {
         if (isset($this->config['resolveType'])) {
-            $fn = $this->config['resolveType'];
-
-            return $fn($objectValue, $context, $info);
+            return ($this->config['resolveType'])($objectValue, $context, $info);
         }
 
         return null;